fix(logs): Make the KeyValuePair per logger
The KeyValuePair defined in the log4j2.xml are defined as static in the
layout.
This leads anyone having defined two appenders with a variable named in
the same way to get always the value defined in the last initialized
appender.

By removing the static modifier for the additional attributes, each
logger now has its own pairs and each appender can use the same key
rather than having to use different ones.